+++ R/grep.R
@@ -1,0 +1,87 @@
+# kate: default-dictionary en_AU
+
+## stringx package for R
+## Copyleft (C) 2021, Marek Gagolewski <https://www.gagolewski.com>
+##
+## This program is free software; you can redistribute it and/or modify
+## it under the terms of the GNU General Public License as published by
+## the Free Software Foundation; either version 2 of the License, or
+## (at your option) any later version.
+##
+## This program is distributed in the hope that it will be useful,
+## but WITHOUT ANY WARRANTY; without even the implied warranty of
+## MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+## GNU General Public License for more details. You have received
+## a copy of the GNU General Public License along with this program.
+
+
+#' @rdname grep
+grep <- function(
+    pattern, x, ignore.case=FALSE, perl=FALSE, value=FALSE,
+    fixed=FALSE, useBytes=FALSE, invert=FALSE
+) {
+
+}
+
+
+#' @rdname grep
+grepl <- function(
+    pattern, x, ignore.case=FALSE, perl=FALSE,
+    fixed=FALSE, useBytes=FALSE
+) {
+
+}
+
+
+#' @rdname grep
+sub <- function(
+    pattern, replacement, x, ignore.case=FALSE, perl=FALSE,
+    fixed=FALSE, useBytes=FALSE
+) {
+
+}
+
+
+#' @rdname grep
+gsub <- function(
+    pattern, replacement, x, ignore.case=FALSE, perl=FALSE,
+    fixed=FALSE, useBytes=FALSE
+) {
+
+}
+
+
+#' @rdname grep
+regexpr <- function(
+    pattern, text, ignore.case=FALSE, perl=FALSE,
+    fixed=FALSE, useBytes=FALSE
+) {
+
+}
+
+
+#' @rdname grep
+gregexpr <- function(
+    pattern, text, ignore.case=FALSE, perl=FALSE,
+    fixed=FALSE, useBytes=FALSE
+) {
+
+}
+
+
+#' @rdname grep
+regexec <- function(
+    pattern, text, ignore.case=FALSE, perl=FALSE,
+    fixed=FALSE, useBytes=FALSE
+) {
+
+}
+
+
+#' @rdname grep
+gregexec <- function(
+    pattern, text, ignore.case=FALSE, perl=FALSE,
+    fixed=FALSE, useBytes=FALSE
+) {
+
+}