+++ R/strwrap.R
@@ -18,17 +18,19 @@
 #' @title
 #' Word-Wrap Text
 #'
+#'
 #' @description
 #' Splits each string into words which are then arranged to form text lines
 #' of no more of a given width.
-#' Might be useful when displaying strings using a monospaced font.
+#'
 #'
 #' @details
+#' Might be useful when displaying strings using a monospaced font.
+#'
+#'
+#' @section Differences from base R:
 #' Replacement for base \code{\link[base]{strwrap}}
 #' implemented with \code{\link[stringi]{stri_wrap}}.
-#'
-#'
-#' Inconsistencies/limitations in/differences from base R:
 #'
 #' \itemize{
 #' \item missing values not propagated