+++ R/substr.R
@@ -19,17 +19,22 @@
 #' Extract or Replace Substrings
 #'
 #' @description
-#' \code{substr} (deprecated synonym: \code{substring}) extracts
+#' \code{substr} extracts
 #' contiguous parts of given character strings.
 #' Its replacement version allows for substituting them with new content.
 #'
 #' @details
+#' \code{substring} is a [DEPRECATED] synonym for \code{substr}.
+#'
+#' Note that these functions can break some meaningful Unicode code point
+#' sequences, e.g., when inputs are not normalised. For extracting
+#' initial parts of strings based on character width, see \code{\link{strtrim}}.
+#'
+#'
+#' @section Differences from base R:
 #' Replacement for base \code{\link[base]{substr}}
 #' and \code{\link[base]{substring}}
 #' implemented with \code{\link[stringi]{stri_sub}}.
-#'
-#'
-#' Inconsistencies in/differences from base R:
 #'
 #' \itemize{
 #' \item \code{substring} is "for compatibility with S", but this should
@@ -76,9 +81,6 @@
 #' The attributes are copied from the longest arguments (similarly
 #' as in the case of binary operators).
 #'
-#' Note that these functions can break some meaningful Unicode code point
-#' sequences, e.g., when inputs are not normalised. For extracting
-#' initial parts of strings based on character width, see \code{\link{strtrim}}.
 #'
 #'
 #' @examples