+++ inst/realtest/startswith-stringx.R
@@ -1,2 +1,15 @@
-E(startsWith("ABABA", c("a", "ab", "aba", "baba", NA), ignore.case=TRUE), c(TRUE, TRUE, TRUE, FALSE, NA))
-E(endsWith("ABABA", c("a", "ab", "aba", "baba", NA), ignore.case=TRUE), c(TRUE, FALSE, TRUE, TRUE, NA))
+x <- c("mario", "Mario", "M\u00E1rio", "M\u00C1RIO", "Mar\u00EDa", "Rosario", NA)
+E(startsWith(x, "mario"), c(TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, NA))
+E(endsWith(x, "mario"), c(TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, NA))
+
+E(startsWith(x, "mario", fixed=NA), c(TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, NA))
+E(endsWith(x, "mario", fixed=NA), c(TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, NA))
+
+E(startsWith(x, "mario", ignore.case=TRUE), c(TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, NA))
+E(endsWith(x, "mario", ignore.case=TRUE), c(TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, NA))
+
+E(startsWith(x, "mario", fixed=NA, ignore.case=TRUE), c(TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, NA))
+E(endsWith(x, "mario", fixed=NA, ignore.case=TRUE), c(TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, NA))
+
+E(startsWith(x, "mario", fixed=NA, strength=1L), c(TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, NA))
+E(endsWith(x, "mario", fixed=NA, strength=1L), c(TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, NA))